refactor(bracket): modularize canvas layout engine
Split layout into bracket-layout/ (types, coords, connectors, labels),
dedupe connector and column math, and route useBracketLayout via
canvasBuilders.
