--- src/lib/tournament/bracket-layout.ts
@@ -1,684 +1,0 @@
-import type { MatchData, MatchStatus } from '@/features/dashboard/types';
-import {
-  CONNECTOR_CORNER_RADIUS,
-  FINALE_LTR_COLUMN_EXTRA,
-  FINAL_FEEDER_EXTRA,
-  MATCH_H,
-  MATCH_ROW_GAP,
-  MATCH_ROW_GAP_COMPACT,
-  MATCH_ROW_GAP_SPACIOUS,
-  MATCH_W,
-  PADDING,
-  ROUND_GAP,
-  ROUND_LABEL_BAND,
-} from '@/config/bracket';
-
-export type BracketCanvasLayout = 'two-sided' | 'one-sided';
-
-export type BracketWing = 'left' | 'right' | 'center';
-
-export interface MatchPosition {
-  x: number;
-  y: number;
-  match: MatchData;
-  wing: BracketWing;
-}
-
-export type BracketLayoutResult = {
-  positions: Array<MatchPosition>;
-  width: number;
-  height: number;
-  layoutMaxRound: number;
-  thirdPlace: MatchData | null;
-};
-
-export function matchWing(
-  round: number,
-  matchIndex: number,
-  maxRound: number
-): BracketWing {
-  if (round === maxRound && matchIndex <= 1) return 'center';
-  const split = 2 ** (maxRound - 1 - round);
-  return matchIndex < split ? 'left' : 'right';
-}
-
-export function resolveBracketLayout(
-  matches: Array<MatchData>,
-  thirdPlaceMatch: boolean
-): {
-  mainMatches: Array<MatchData>;
-  thirdPlace: MatchData | null;
-  layoutMaxRound: number;
-} {
-  if (matches.length === 0) {
-    return { mainMatches: [], thirdPlace: null, layoutMaxRound: 0 };
-  }
-  const dataMaxRound = Math.max(...matches.map((m) => m.round));
-
-  if (!thirdPlaceMatch) {
-    return {
-      mainMatches: matches,
-      thirdPlace: null,
-      layoutMaxRound: dataMaxRound,
-    };
-  }
-
-  const sameRoundThird = matches.find(
-    (m) => m.round === dataMaxRound && m.matchIndex === 1
-  );
-  if (sameRoundThird) {
-    return {
-      mainMatches: matches.filter(
-        (m) => !(m.round === dataMaxRound && m.matchIndex === 1)
-      ),
-      thirdPlace: sameRoundThird,
-      layoutMaxRound: dataMaxRound,
-    };
-  }
-
-  const extraRoundThird = matches.find(
-    (m) => m.round === dataMaxRound && m.matchIndex === 0
-  );
-  if (
-    extraRoundThird &&
-    !matches.some((m) => m.round === dataMaxRound && m.matchIndex === 1)
-  ) {
-    return {
-      mainMatches: matches.filter((m) => m.round !== dataMaxRound),
-      thirdPlace: extraRoundThird,
-      layoutMaxRound: dataMaxRound - 1,
-    };
-  }
-
-  return {
-    mainMatches: matches,
-    thirdPlace: null,
-    layoutMaxRound: dataMaxRound,
-  };
-}
-
-export function isBracketFinal(
-  match: MatchData,
-  layoutMaxRound: number
-): boolean {
-  return match.round === layoutMaxRound && match.matchIndex === 0;
-}
-
-export function layoutCenterX(maxRound: number): number {
-  return PADDING + maxRound * ROUND_GAP + MATCH_W / 2;
-}
-
-/** Two-sided wing vertical gap: tighter when bracket has more than three rounds. */
-export function twoSidedMatchRowGap(layoutMaxRound: number): number {
-  return layoutMaxRound > 2 ? MATCH_ROW_GAP_SPACIOUS : MATCH_ROW_GAP_COMPACT;
-}
-
-/** Layout left edge sits inside the scaled final box (`scale-120`, center origin). */
-const FINAL_SCALE_OVERHANG = (MATCH_W * (1.2 - 1)) / 2;
-
-function roundColumnOffset(
-  round: number,
-  wing: BracketWing,
-  layoutMaxRound: number
-): number {
-  let offset = (layoutMaxRound - round) * ROUND_GAP;
-  if (wing !== 'center' && round < layoutMaxRound) {
-    offset += FINAL_FEEDER_EXTRA;
-  }
-  return offset;
-}
-
-export function roundLabelCenterX(
-  round: number,
-  wing: BracketWing,
-  layoutMaxRound: number
-): number {
-  const centerX = layoutCenterX(layoutMaxRound);
-  const offset = roundColumnOffset(round, wing, layoutMaxRound);
-  if (wing === 'center') return centerX;
-  if (wing === 'left') return centerX - offset;
-  return centerX + offset;
-}
-
-function matchX(
-  round: number,
-  matchIndex: number,
-  maxRound: number,
-  centerX: number
-): number {
-  const wing = matchWing(round, matchIndex, maxRound);
-  const offset = roundColumnOffset(round, wing, maxRound);
-  if (wing === 'center') return centerX - MATCH_W / 2;
-  if (wing === 'left') return centerX - MATCH_W / 2 - offset;
-  return centerX + MATCH_W / 2 + offset - MATCH_W;
-}
-
-export function buildTwoSidedLayout(
-  matches: Array<MatchData>,
-  thirdPlaceMatch: boolean
-) {
-  const { mainMatches, thirdPlace, layoutMaxRound } = resolveBracketLayout(
-    matches,
-    thirdPlaceMatch
-  );
-  const maxRound = layoutMaxRound;
-
-  if (mainMatches.length === 0) {
-    return {
-      positions: [],
-      width: 0,
-      height: 0,
-      layoutMaxRound,
-      thirdPlace,
-    };
-  }
-
-  const rounds = new Map<number, Array<MatchData>>();
-  for (const m of mainMatches) {
-    const arr = rounds.get(m.round) ?? [];
-    arr.push(m);
-    rounds.set(m.round, arr);
-  }
-  for (const arr of rounds.values()) {
-    arr.sort((a, b) => a.matchIndex - b.matchIndex);
-  }
-
-  const roundNums = Array.from(rounds.keys()).sort((a, b) => a - b);
-  const r0Count = rounds.get(roundNums[0]!)?.length ?? 1;
-  const rowGap = twoSidedMatchRowGap(maxRound);
-  const totalMainHeight = r0Count * MATCH_H + (r0Count - 1) * rowGap;
-
-  const centerX = layoutCenterX(maxRound);
-  const matchTop = PADDING + ROUND_LABEL_BAND;
-  const yByKey = new Map<string, number>();
-
-  function layoutWingTree(wing: 'left' | 'right') {
-    const r0InWing = (rounds.get(0) ?? []).filter(
-      (m) => matchWing(0, m.matchIndex, maxRound) === wing
-    );
-    if (r0InWing.length === 0) return;
-
-    const wingHeight =
-      r0InWing.length * MATCH_H + (r0InWing.length - 1) * rowGap;
-    const slotH = wingHeight / r0InWing.length;
-
-    for (let i = 0; i < r0InWing.length; i++) {
-      const match = r0InWing[i]!;
-      yByKey.set(
-        `${match.round}-${match.matchIndex}`,
-        matchTop + i * slotH + (slotH - MATCH_H) / 2
-      );
-    }
-
-    for (const round of roundNums.slice(1)) {
-      if (round === maxRound) continue;
-      for (const match of rounds.get(round) ?? []) {
-        if (matchWing(match.round, match.matchIndex, maxRound) !== wing) {
-          continue;
-        }
-        const childA = yByKey.get(`${round - 1}-${match.matchIndex * 2}`);
-        const childB = yByKey.get(`${round - 1}-${match.matchIndex * 2 + 1}`);
-        const midA =
-          childA != null ? childA + MATCH_H / 2 : matchTop + wingHeight / 2;
-        const midB =
-          childB != null ? childB + MATCH_H / 2 : matchTop + wingHeight / 2;
-        yByKey.set(
-          `${match.round}-${match.matchIndex}`,
-          (midA + midB) / 2 - MATCH_H / 2
-        );
-      }
-    }
-  }
-
-  layoutWingTree('left');
-  layoutWingTree('right');
-
-  const leftSf = yByKey.get(`${maxRound - 1}-0`);
-  const rightSf = yByKey.get(`${maxRound - 1}-1`);
-  const finalY =
-    leftSf != null && rightSf != null
-      ? (leftSf + rightSf) / 2
-      : matchTop + totalMainHeight / 2 - MATCH_H / 2;
-  yByKey.set(`${maxRound}-0`, finalY);
-
-  const positions: Array<MatchPosition> = [];
-  for (const round of roundNums) {
-    for (const match of rounds.get(round)!) {
-      const y = yByKey.get(`${match.round}-${match.matchIndex}`)!;
-      positions.push({
-        x: matchX(match.round, match.matchIndex, maxRound, centerX),
-        y,
-        match,
-        wing: matchWing(match.round, match.matchIndex, maxRound),
-      });
-    }
-  }
-
-  if (thirdPlace) {
-    const finalPos = positions.find(
-      (p) => p.match.round === maxRound && p.match.matchIndex === 0
-    );
-    const x = finalPos?.x ?? centerX - MATCH_W / 2;
-    const y = matchTop + totalMainHeight + PADDING + MATCH_H;
-    positions.push({
-      x,
-      y,
-      match: thirdPlace,
-      wing: 'center',
-    });
-  }
-
-  const maxX = Math.max(...positions.map((p) => p.x)) + MATCH_W + PADDING;
-  const maxY = Math.max(...positions.map((p) => p.y)) + MATCH_H + PADDING;
-
-  return {
-    positions,
-    width: maxX,
-    height: maxY,
-    layoutMaxRound,
-    thirdPlace,
-  };
-}
-
-function oneSidedMatchX(round: number, layoutMaxRound: number): number {
-  let x = PADDING + round * ROUND_GAP;
-  if (round === layoutMaxRound && layoutMaxRound > 0) {
-    x += FINALE_LTR_COLUMN_EXTRA;
-  }
-  return x;
-}
-
-export function oneSidedRoundLabelX(
-  round: number,
-  layoutMaxRound: number
-): number {
-  return oneSidedMatchX(round, layoutMaxRound) + MATCH_W / 2;
-}
-
-export function buildOneSidedLayout(
-  matches: Array<MatchData>,
-  thirdPlaceMatch: boolean
-): BracketLayoutResult {
-  const { mainMatches, thirdPlace, layoutMaxRound } = resolveBracketLayout(
-    matches,
-    thirdPlaceMatch
-  );
-  const maxRound = layoutMaxRound;
-
-  if (mainMatches.length === 0) {
-    return {
-      positions: [],
-      width: 0,
-      height: 0,
-      layoutMaxRound,
-      thirdPlace,
-    };
-  }
-
-  const rounds = new Map<number, Array<MatchData>>();
-  for (const m of mainMatches) {
-    const arr = rounds.get(m.round) ?? [];
-    arr.push(m);
-    rounds.set(m.round, arr);
-  }
-  for (const arr of rounds.values()) {
-    arr.sort((a, b) => a.matchIndex - b.matchIndex);
-  }
-
-  const roundNums = Array.from(rounds.keys()).sort((a, b) => a - b);
-  const r0Matches = rounds.get(roundNums[0]!) ?? [];
-  const r0Count = r0Matches.length;
-  const totalMainHeight =
-    r0Count * MATCH_H + Math.max(0, r0Count - 1) * MATCH_ROW_GAP;
-
-  const matchTop = PADDING + ROUND_LABEL_BAND;
-  const yByKey = new Map<string, number>();
-  const slotH = r0Count > 0 ? totalMainHeight / r0Count : MATCH_H;
-
-  for (let i = 0; i < r0Matches.length; i++) {
-    const match = r0Matches[i]!;
-    yByKey.set(
-      `${match.round}-${match.matchIndex}`,
-      matchTop + i * slotH + (slotH - MATCH_H) / 2
-    );
-  }
-
-  for (const round of roundNums.slice(1)) {
-    for (const match of rounds.get(round) ?? []) {
-      const childA = yByKey.get(`${round - 1}-${match.matchIndex * 2}`);
-      const childB = yByKey.get(`${round - 1}-${match.matchIndex * 2 + 1}`);
-      const midA =
-        childA != null ? childA + MATCH_H / 2 : matchTop + totalMainHeight / 2;
-      const midB =
-        childB != null ? childB + MATCH_H / 2 : matchTop + totalMainHeight / 2;
-      yByKey.set(
-        `${match.round}-${match.matchIndex}`,
-        (midA + midB) / 2 - MATCH_H / 2
-      );
-    }
-  }
-
-  const positions: Array<MatchPosition> = [];
-  for (const round of roundNums) {
-    for (const match of rounds.get(round)!) {
-      positions.push({
-        x: oneSidedMatchX(match.round, maxRound),
-        y: yByKey.get(`${match.round}-${match.matchIndex}`)!,
-        match,
-        wing: 'left',
-      });
-    }
-  }
-
-  if (thirdPlace) {
-    const finalPos = positions.find(
-      (p) => p.match.round === maxRound && p.match.matchIndex === 0
-    );
-    const x = finalPos?.x ?? oneSidedMatchX(maxRound, maxRound);
-    const y = matchTop + totalMainHeight + PADDING + MATCH_H;
-    positions.push({
-      x,
-      y,
-      match: thirdPlace,
-      wing: 'center',
-    });
-  }
-
-  const maxX = Math.max(...positions.map((p) => p.x)) + MATCH_W + PADDING;
-  const maxY = Math.max(...positions.map((p) => p.y)) + MATCH_H + PADDING;
-
-  return {
-    positions,
-    width: maxX,
-    height: maxY,
-    layoutMaxRound,
-    thirdPlace,
-  };
-}
-
-export interface BracketConnectorPath {
-  d: string;
-  status: MatchStatus;
-}
-
-/** Child → stub (LTR wing): horizontal from child right, fillet, vertical to bus. */
-export function buildConnectorChildLeg(
-  childRightX: number,
-  childMidY: number,
-  midX: number,
-  parentMidY: number,
-  radius: number = CONNECTOR_CORNER_RADIUS
-): string {
-  const dy = Math.abs(parentMidY - childMidY);
-  const dx = midX - childRightX;
-  if (dy < 0.5 || dx <= 0) {
-    return `M ${childRightX} ${childMidY} L ${midX} ${parentMidY}`;
-  }
-  const r = Math.min(
-    radius,
-    dx * 0.45,
-    dy * 0.45,
-    dx / 2 - 0.25,
-    dy / 2 - 0.25
-  );
-  const rr = Math.max(2, r);
-  const down = parentMidY > childMidY;
-
-  if (down) {
-    return [
-      `M ${childRightX} ${childMidY}`,
-      `L ${midX - rr} ${childMidY}`,
-      `A ${rr} ${rr} 0 0 1 ${midX} ${childMidY + rr}`,
-      `L ${midX} ${parentMidY}`,
-    ].join(' ');
-  }
-
-  return [
-    `M ${childRightX} ${childMidY}`,
-    `L ${midX - rr} ${childMidY}`,
-    `A ${rr} ${rr} 0 0 0 ${midX} ${childMidY - rr}`,
-    `L ${midX} ${parentMidY}`,
-  ].join(' ');
-}
-
-/** Child → stub (RTL wing): horizontal from child left, fillet, vertical to bus. */
-export function buildConnectorChildLegRtl(
-  childLeftX: number,
-  childMidY: number,
-  midX: number,
-  parentMidY: number,
-  radius: number = CONNECTOR_CORNER_RADIUS
-): string {
-  const dy = Math.abs(parentMidY - childMidY);
-  const dx = childLeftX - midX;
-  if (dy < 0.5 || dx <= 0) {
-    return `M ${childLeftX} ${childMidY} L ${midX} ${parentMidY}`;
-  }
-  const r = Math.min(
-    radius,
-    dx * 0.45,
-    dy * 0.45,
-    dx / 2 - 0.25,
-    dy / 2 - 0.25
-  );
-  const rr = Math.max(2, r);
-  const down = parentMidY > childMidY;
-
-  if (down) {
-    return [
-      `M ${childLeftX} ${childMidY}`,
-      `L ${midX + rr} ${childMidY}`,
-      `A ${rr} ${rr} 0 0 0 ${midX} ${childMidY + rr}`,
-      `L ${midX} ${parentMidY}`,
-    ].join(' ');
-  }
-
-  return [
-    `M ${childLeftX} ${childMidY}`,
-    `L ${midX + rr} ${childMidY}`,
-    `A ${rr} ${rr} 0 0 1 ${midX} ${childMidY - rr}`,
-    `L ${midX} ${parentMidY}`,
-  ].join(' ');
-}
-
-/** Shared horizontal from stub column into the parent match (LTR). */
-export function buildConnectorTrunk(
-  midX: number,
-  parentMidY: number,
-  parentLeftX: number
-): string {
-  return `M ${midX} ${parentMidY} L ${parentLeftX} ${parentMidY}`;
-}
-
-/** Shared horizontal from stub column into the parent match (RTL). */
-export function buildConnectorTrunkRtl(
-  midX: number,
-  parentMidY: number,
-  parentRightX: number
-): string {
-  return `M ${midX} ${parentMidY} L ${parentRightX} ${parentMidY}`;
-}
-
-const connectorStub = (ROUND_GAP - MATCH_W) / 2;
-
-type ConnectorDirection = 'ltr' | 'rtl';
-
-function pushConnector(
-  paths: Array<BracketConnectorPath>,
-  d: string,
-  status: MatchStatus
-) {
-  paths.push({ d, status });
-}
-
-function appendLtrToParent(
-  paths: Array<BracketConnectorPath>,
-  parent: MatchPosition,
-  children: Array<MatchPosition | undefined>,
-  trunkEndX?: number
-) {
-  const parentMidY = parent.y + MATCH_H / 2;
-  const status = parent.match.status;
-  const parentLeftX = parent.x;
-  const midX = parentLeftX - connectorStub;
-  const trunkX = trunkEndX ?? parentLeftX;
-  let hasChild = false;
-  for (const child of children) {
-    if (!child) continue;
-    hasChild = true;
-    pushConnector(
-      paths,
-      buildConnectorChildLeg(
-        child.x + MATCH_W,
-        child.y + MATCH_H / 2,
-        midX,
-        parentMidY
-      ),
-      status
-    );
-  }
-  if (hasChild) {
-    pushConnector(paths, buildConnectorTrunk(midX, parentMidY, trunkX), status);
-  }
-}
-
-function appendOneSidedFinalFeeders(
-  paths: Array<BracketConnectorPath>,
-  parent: MatchPosition,
-  childA: MatchPosition | undefined,
-  childB: MatchPosition | undefined
-) {
-  const feeders = [childA, childB].filter((c): c is MatchPosition => c != null);
-  if (feeders.length === 0) return;
-
-  const parentMidY = parent.y + MATCH_H / 2;
-  const status = parent.match.status;
-  const attachX = parent.x - FINAL_SCALE_OVERHANG;
-  const maxChildRight = Math.max(...feeders.map((c) => c.x + MATCH_W));
-  const minTrunk = FINALE_LTR_COLUMN_EXTRA - FINAL_SCALE_OVERHANG;
-  const busX = Math.max(maxChildRight + connectorStub, attachX - minTrunk);
-
-  for (const child of feeders) {
-    pushConnector(
-      paths,
-      buildConnectorChildLeg(
-        child.x + MATCH_W,
-        child.y + MATCH_H / 2,
-        busX,
-        parentMidY
-      ),
-      status
-    );
-  }
-  pushConnector(paths, buildConnectorTrunk(busX, parentMidY, attachX), status);
-}
-
-function appendToParent(
-  paths: Array<BracketConnectorPath>,
-  parent: MatchPosition,
-  direction: ConnectorDirection,
-  children: Array<MatchPosition | undefined>
-) {
-  const parentMidY = parent.y + MATCH_H / 2;
-  const status = parent.match.status;
-
-  if (direction === 'ltr') {
-    appendLtrToParent(paths, parent, children);
-    return;
-  }
-
-  const parentRightX = parent.x + MATCH_W;
-  const midX = parentRightX + connectorStub;
-  let hasChild = false;
-  for (const child of children) {
-    if (!child) continue;
-    hasChild = true;
-    pushConnector(
-      paths,
-      buildConnectorChildLegRtl(
-        child.x,
-        child.y + MATCH_H / 2,
-        midX,
-        parentMidY
-      ),
-      status
-    );
-  }
-  if (hasChild) {
-    pushConnector(
-      paths,
-      buildConnectorTrunkRtl(midX, parentMidY, parentRightX),
-      status
-    );
-  }
-}
-
-export function buildTwoSidedConnectors(
-  positions: Array<MatchPosition>,
-  layoutMaxRound: number
-) {
-  const paths: Array<BracketConnectorPath> = [];
-  const posMap = new Map<string, MatchPosition>();
-  for (const p of positions) {
-    posMap.set(`${p.match.round}-${p.match.matchIndex}`, p);
-  }
-
-  for (const pos of positions) {
-    if (pos.match.round === 0) continue;
-
-    const childA = posMap.get(
-      `${pos.match.round - 1}-${pos.match.matchIndex * 2}`
-    );
-    const childB = posMap.get(
-      `${pos.match.round - 1}-${pos.match.matchIndex * 2 + 1}`
-    );
-
-    if (isBracketFinal(pos.match, layoutMaxRound)) {
-      appendToParent(paths, pos, 'ltr', [childA]);
-      appendToParent(paths, pos, 'rtl', [childB]);
-      continue;
-    }
-
-    if (pos.wing === 'left') {
-      appendToParent(paths, pos, 'ltr', [childA, childB]);
-      continue;
-    }
-
-    if (pos.wing === 'right') {
-      appendToParent(paths, pos, 'rtl', [childA, childB]);
-    }
-  }
-
-  return paths;
-}
-
-export function buildOneSidedConnectors(
-  positions: Array<MatchPosition>,
-  layoutMaxRound: number
-) {
-  const paths: Array<BracketConnectorPath> = [];
-  const posMap = new Map<string, MatchPosition>();
-  for (const p of positions) {
-    posMap.set(`${p.match.round}-${p.match.matchIndex}`, p);
-  }
-
-  for (const pos of positions) {
-    if (pos.match.round === 0) continue;
-
-    const childA = posMap.get(
-      `${pos.match.round - 1}-${pos.match.matchIndex * 2}`
-    );
-    const childB = posMap.get(
-      `${pos.match.round - 1}-${pos.match.matchIndex * 2 + 1}`
-    );
-
-    if (isBracketFinal(pos.match, layoutMaxRound)) {
-      appendOneSidedFinalFeeders(paths, pos, childA, childB);
-      continue;
-    }
-
-    appendLtrToParent(paths, pos, [childA, childB]);
-  }
-
-  return paths;
-}
